requirements.txt
certifi==2018.1.18
chardet==3.0.4
Django==1.11.4
django-jet==1.0.6
django-model-utils==3.0.0
django-sendsms==0.2.3
idna==2.6
jsonfield==2.0.2
olefile==0.44
Pillow==4.2.1
psycopg2==2.7.3
PyJWT==1.5.3
PySocks==1.6.8
pytz==2017.2
raven==6.3.0
requests==2.18.4
singlemodeladmin==0.8
six==1.11.0
twilio==6.10.1
urllib3==1.22
